refactoring to repository, making more modular
