+++ lib/data/repository.dart
@@ -1,0 +1,29 @@
+import 'package:test_app/database.dart';
+
+
+
+class Repository {
+
+  static final Repository _repo = new Repository._internal();
+
+  BookDatabase database;
+
+  static Repository get() {
+    return _repo;
+  }
+
+  Repository._internal() {
+    database = BookDatabase.get();
+  }
+
+
+
+
+
+
+
+
+
+
+
+}
